move get_current_site to utils
